+++ src/main/java/com/travel/role/domain/user/dto/auth/LoginRequestDTO.java
@@ -1,4 +1,4 @@
-package com.travel.role.domain.user.dto;
+package com.travel.role.domain.user.dto.auth;
 
 import lombok.Data;
