+++ src/main/java/com/travel/role/domain/user/dto/auth/SignUpRequestDTO.java
@@ -1,11 +1,8 @@
-package com.travel.role.domain.user.dto;
+package com.travel.role.domain.user.dto.auth;
 
 import java.time.LocalDate;
-import java.util.Date;
 
 import org.springframework.format.annotation.DateTimeFormat;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
 
 import lombok.Data;
